+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Dnsmasq Local Cookbook CHANGELOG
 ================================
 
+Unreleased
+----------
+
 v0.3.0 (2016-05-26)
 -------------------
 - Fix custom config properties/attributes under Chef 11