+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 Dnsmasq Local Cookbook CHANGELOG
 ================================
 
-Unreleased
-----------
+v0.4.0 (2016-06-24)
+-------------------
 - Take over management of the Dnsmasq environment variables
 - Add a warning comment to all Chef-managed config files